Typecheck DCLLoss
