+++ pyproject.toml
@@ -196,7 +196,6 @@ exclude = '''(?x)(
     lightly/loss/dino_loss.py |
     lightly/loss/sym_neg_cos_sim_loss.py |
     lightly/loss/vicregl_loss.py |
-    lightly/loss/dcl_loss.py |
     lightly/loss/regularizer/co2.py |
     lightly/loss/barlow_twins_loss.py |
     lightly/data/dataset.py |
@@ -246,7 +245,6 @@ exclude = '''(?x)(
     tests/loss/test_DINOLoss.py |
     tests/loss/test_VICRegLLoss.py |
     tests/loss/test_CO2Regularizer.py |
-    tests/loss/test_DCLLoss.py |
     tests/loss/test_barlow_twins_loss.py |
     tests/loss/test_SymNegCosineSimilarityLoss.py |
     tests/loss/test_MemoryBank.py |
